+++ nym-vpn-android/app/src/main/java/net/nymtech/nymvpn/manager/environment/NymEnvironmentManager.kt
@@ -7,8 +7,6 @@ import javax.inject.Inject
 
 class NymEnvironmentManager @Inject constructor(private val backendManager: BackendManager) : EnvironmentManager {
 
-	override suspend fun isMixnetTuningEnabled(): Boolean = getFeatureFlags()?.isMixnetTuningEnabled() ?: false
-
 	override suspend fun isDomainFrontingEnabled(): Boolean = getFeatureFlags()?.isDomainFrontingEnabled() ?: false
 
 	private suspend fun getFeatureFlags(): FeatureFlags? = try {