+++ nym-vpn-app/CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Allow adding custom apps (user selected) to split tunneling for Windows & Linux
+- Mixnet tuning settings
+
+### CHANGED
+
+- Switch to daily allowance usage from monthly
 
 ## [1.30.0] - 2026-05-29